Use a pure FFI type for ompPopCount
`ompPopCount` previously used a foreign call with a result in `IO`, wrapped in `unsafeDupablePerformIO`. Use a pure FFI call instead. Additionally, the `Int` argument to this function was previously unboxed manually but its result wasn't. I'm not sure how much difference it makes in practice, but uniformity is nice; unbox the result too.